+++ pkg/kubernetesutils/kubeconfigutils/KubeConfig_test.go
@@ -1,10 +1,11 @@
-package kubeconfig
+package kubeconfigutils_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
 	"github.com/asciich/asciichgolangpublic/files"
+	"github.com/asciich/asciichgolangpublic/pkg/kubernetesutils/kubeconfigutils"
 	"github.com/asciich/asciichgolangpublic/pkg/mustutils"
 	"github.com/asciich/asciichgolangpublic/testutils"
 )
@@ -24,7 +25,7 @@ func Test_GetUserEntryByUserName(t *testing.T) {
 		t.Run(
 			testutils.MustFormatAsTestname(tt),
 			func(t *testing.T) {
-				kubeConfig, err := LoadFromFilePath(tt.path, true)
+				kubeConfig, err := kubeconfigutils.LoadFromFilePath(tt.path, true)
 				require.NoError(t, err)
 
 				entry, err := kubeConfig.GetUserEntryByName(tt.userName)
@@ -35,7 +36,7 @@ func Test_GetUserEntryByUserName(t *testing.T) {
 	}
 
 	t.Run("Unknown user name", func(t *testing.T) {
-		kubeConfig, err := LoadFromFilePath("./testdata/cluster-c.yaml", true)
+		kubeConfig, err := kubeconfigutils.LoadFromFilePath("./testdata/cluster-c.yaml", true)
 		require.NoError(t, err)
 
 		entry, err := kubeConfig.GetUserEntryByName("this-user-does-not-exist")
@@ -60,7 +61,7 @@ func Test_GetUserNameByContextName(t *testing.T) {
 		t.Run(
 			testutils.MustFormatAsTestname(tt),
 			func(t *testing.T) {
-				kubeConfig, err := LoadFromFilePath(tt.path, true)
+				kubeConfig, err := kubeconfigutils.LoadFromFilePath(tt.path, true)
 				require.NoError(t, err)
 
 				userName, err := kubeConfig.GetUserNameByContextName(tt.contextName)
@@ -71,7 +72,7 @@ func Test_GetUserNameByContextName(t *testing.T) {
 	}
 
 	t.Run("Unknown context name", func(t *testing.T) {
-		kubeConfig, err := LoadFromFilePath("./testdata/cluster-c.yaml", true)
+		kubeConfig, err := kubeconfigutils.LoadFromFilePath("./testdata/cluster-c.yaml", true)
 		require.NoError(t, err)
 
 		entry, err := kubeConfig.GetUserNameByContextName("this-context-does-not-exist")
@@ -96,7 +97,7 @@ func TestKubeConfig_LoadFromPath(t *testing.T) {
 		t.Run(
 			testutils.MustFormatAsTestname(tt),
 			func(t *testing.T) {
-				kubeConfig, err := LoadFromFilePath(tt.path, true)
+				kubeConfig, err := kubeconfigutils.LoadFromFilePath(tt.path, true)
 				require.NoError(t, err)
 
 				require.EqualValues(
@@ -130,18 +131,18 @@ func TestKubeConfig_IsLoadableByKubectl(t *testing.T) {
 			func(t *testing.T) {
 				const verbose bool = true
 
-				isLoadable, err := IsFilePathLoadableByKubectl(tt.path, verbose)
+				isLoadable, err := kubeconfigutils.IsFilePathLoadableByKubectl(tt.path, verbose)
 				require.NoError(t, err)
 				require.True(t, isLoadable)
 
-				kubeConfig, err := LoadFromFilePath(tt.path, verbose)
+				kubeConfig, err := kubeconfigutils.LoadFromFilePath(tt.path, verbose)
 				require.NoError(t, err)
 
 				tempFilePath, err := kubeConfig.WriteToTemporaryFileAndGetPath(verbose)
 				require.NoError(t, err)
 				defer files.MustDeleteFileByPath(tempFilePath, verbose)
 
-				isLoadable, err = IsFilePathLoadableByKubectl(tt.path, verbose)
+				isLoadable, err = kubeconfigutils.IsFilePathLoadableByKubectl(tt.path, verbose)
 				require.NoError(t, err)
 				require.True(t, isLoadable)
 
@@ -166,16 +167,16 @@ func TestKubeConfig_CheckContextsUsingKubectl(t *testing.T) {
 			func(t *testing.T) {
 				const verbose bool = true
 
-				require.EqualValues(t, tt.expectedContextNames, mustutils.Must(ListContextNamesUsingKubectl(tt.path, verbose)))
+				require.EqualValues(t, tt.expectedContextNames, mustutils.Must(kubeconfigutils.ListContextNamesUsingKubectl(tt.path, verbose)))
 
-				kubeConfig, err := LoadFromFilePath(tt.path, verbose)
+				kubeConfig, err := kubeconfigutils.LoadFromFilePath(tt.path, verbose)
 				require.NoError(t, err)
 
 				tempFilePath, err := kubeConfig.WriteToTemporaryFileAndGetPath(verbose)
 				require.NoError(t, err)
 				defer files.MustDeleteFileByPath(tempFilePath, verbose)
 
-				require.EqualValues(t, tt.expectedContextNames, mustutils.Must(ListContextNamesUsingKubectl(tempFilePath, verbose)))
+				require.EqualValues(t, tt.expectedContextNames, mustutils.Must(kubeconfigutils.ListContextNamesUsingKubectl(tempFilePath, verbose)))
 
 			},
 		)
@@ -202,13 +203,13 @@ func TestKubeConfig_MergeTwoConfigs(t *testing.T) {
 			func(t *testing.T) {
 				const verbose = true
 
-				kubeConfig1, err := LoadFromFilePath(tt.path1, true)
+				kubeConfig1, err := kubeconfigutils.LoadFromFilePath(tt.path1, true)
 				require.NoError(t, err)
 
-				kubeConfig2, err := LoadFromFilePath(tt.path2, true)
+				kubeConfig2, err := kubeconfigutils.LoadFromFilePath(tt.path2, true)
 				require.NoError(t, err)
 
-				merged, err := MergeConfig(kubeConfig1, kubeConfig2)
+				merged, err := kubeconfigutils.MergeConfig(kubeConfig1, kubeConfig2)
 				require.NoError(t, err)
 
 				require.EqualValues(t, tt.expectedNames, mustutils.Must(merged.GetClusterNames()))
@@ -218,7 +219,7 @@ func TestKubeConfig_MergeTwoConfigs(t *testing.T) {
 
 				defer files.MustDeleteFileByPath(tempFilePath, verbose)
 
-				require.EqualValues(t, tt.expectedNames, mustutils.Must(ListContextNamesUsingKubectl(tempFilePath, verbose)))
+				require.EqualValues(t, tt.expectedNames, mustutils.Must(kubeconfigutils.ListContextNamesUsingKubectl(tempFilePath, verbose)))
 			},
 		)
 	}
@@ -227,21 +228,21 @@ func TestKubeConfig_MergeTwoConfigs(t *testing.T) {
 func TestKubeConfig_MergeThreeConfigs(t *testing.T) {
 	const verbose = true
 
-	kubeConfig1, err := LoadFromFilePath("./testdata/cluster-a.yaml", true)
+	kubeConfig1, err := kubeconfigutils.LoadFromFilePath("./testdata/cluster-a.yaml", true)
 	require.NoError(t, err)
 
-	kubeConfig2, err := LoadFromFilePath("./testdata/cluster-b.yaml", true)
+	kubeConfig2, err := kubeconfigutils.LoadFromFilePath("./testdata/cluster-b.yaml", true)
 	require.NoError(t, err)
 
-	kubeConfig3, err := LoadFromFilePath("./testdata/cluster-c.yaml", true)
+	kubeConfig3, err := kubeconfigutils.LoadFromFilePath("./testdata/cluster-c.yaml", true)
 	require.NoError(t, err)
 
-	merged1, err := MergeConfig(kubeConfig2, kubeConfig3)
+	merged1, err := kubeconfigutils.MergeConfig(kubeConfig2, kubeConfig3)
 	require.NoError(t, err)
 
 	require.EqualValues(t, []string{"kind-cluster-b", "kind-cluster-c"}, mustutils.Must(merged1.GetClusterNames()))
 
-	merged2, err := MergeConfig(merged1, kubeConfig2, kubeConfig3, kubeConfig1)
+	merged2, err := kubeconfigutils.MergeConfig(merged1, kubeConfig2, kubeConfig3, kubeConfig1)
 	require.NoError(t, err)
 
 	require.EqualValues(
@@ -257,6 +258,6 @@ func TestKubeConfig_MergeThreeConfigs(t *testing.T) {
 	require.EqualValues(
 		t,
 		[]string{"kind-cluster-a", "kind-cluster-b", "kind-cluster-c"},
-		mustutils.Must(ListContextNamesUsingKubectl(tempFilePath, verbose)),
+		mustutils.Must(kubeconfigutils.ListContextNamesUsingKubectl(tempFilePath, verbose)),
 	)
 }
