+++ pkg/kubernetesutils/kubeconfigutils/testdata/cluster-b.yaml
