+++ pkg/kubernetesutils/kubeconfigutils/testdata/cluster-c.yaml
